+++ server/api/controllers/task-lists/delete.js
@@ -93,6 +93,11 @@ module.exports = {
       throw Errors.NOT_ENOUGH_RIGHTS;
     }
 
+    // A locked card is read-only for everyone except instance admins.
+    if (card && card.isLocked && currentUser.role !== User.Roles.ADMIN) {
+      throw Errors.NOT_ENOUGH_RIGHTS;
+    }
+
     taskList = await sails.helpers.taskLists.deleteOne.with({
       project,
       board,
